removed normal server error test for jaxrs as it does not support return a type of response
